requirements.txt
web.py==0.37
PyYAML==3.11
